Refactor logic module structure and update imports for hdsemg-select
